(GH-305) update exitcodes to long

+++ src/chocolatey/infrastructure.app/domain/installers/IInstaller.cs
@@ -30,8 +30,8 @@ namespace chocolatey.infrastructure.app.domain.installers
         string UninstallExecutable { get; }
         string SilentUninstall { get; }
         string OtherUninstallOptions { get; }
-        IEnumerable<int> ValidInstallExitCodes { get; }
-        IEnumerable<int> ValidUninstallExitCodes { get; }
+        IEnumerable<long> ValidInstallExitCodes { get; }
+        IEnumerable<long> ValidUninstallExitCodes { get; }
 
         string build_install_command_arguments(bool logFile, bool customInstallLocation, bool languageRequested);
         string build_uninstall_command_arguments();
