+++ src/chocolatey/infrastructure.app/domain/installers/IzPackInstaller.cs
@@ -39,11 +39,11 @@ namespace chocolatey.infrastructure.app.domain.installers
             UninstallExecutable = "java"; //currently untested
             SilentUninstall = "-jar \"{0}\"".format_with(InstallTokens.UNINSTALLER_LOCATION);
             OtherUninstallOptions = "";
-            ValidInstallExitCodes = new List<int>
+            ValidInstallExitCodes = new List<long>
             {
                 0
             };
-            ValidUninstallExitCodes = new List<int>
+            ValidUninstallExitCodes = new List<long>
             {
                 0
             };
