+++ src/chocolatey/infrastructure.app/domain/installers/PackageForTheWebInstaller.cs
@@ -36,11 +36,11 @@ namespace chocolatey.infrastructure.app.domain.installers
             UninstallExecutable = "\"{0}\"".format_with(InstallTokens.UNINSTALLER_LOCATION);
             SilentUninstall = "/s";
             OtherUninstallOptions = "";
-            ValidInstallExitCodes = new List<int>
+            ValidInstallExitCodes = new List<long>
             {
                 0
             };
-            ValidUninstallExitCodes = new List<int>
+            ValidUninstallExitCodes = new List<long>
             {
                 0
             };
