+++ src/chocolatey/infrastructure.app/domain/installers/SetupFactoryInstaller.cs
@@ -45,11 +45,11 @@ namespace chocolatey.infrastructure.app.domain.installers
             UninstallExecutable = "\"{0}\"".format_with(InstallTokens.UNINSTALLER_LOCATION);
             SilentUninstall = "/S";
             OtherUninstallOptions = "";
-            ValidInstallExitCodes = new List<int>
+            ValidInstallExitCodes = new List<long>
             {
                 0
             };
-            ValidUninstallExitCodes = new List<int>
+            ValidUninstallExitCodes = new List<long>
             {
                 0
             };
